chore: applied code standard
